Finished criar, restaurar e gravar butttons. Refactored exceptions

+++ project_java/src/main/java/li3/grupo54/View/Queries/Query3View.java
@@ -76,11 +76,11 @@ public class Query3View implements IQueryViewFX {
     Query3Results res = (Query3Results) results;
     double time = Crono.stop();
 
-    List<MyTriple<Integer, Integer, Float>> res1 = ((Query3Results) results).getResults();
+    List<MyTriple<Integer, Integer, Double>> res1 = ((Query3Results) results).getResults();
 
-    List<MyFour<Month, Integer, Integer, Float>> apresentar = new ArrayList<>(12);
+    List<MyFour<Month, Integer, Integer, Double>> apresentar = new ArrayList<>(12);
     int i = 1;
-    for (MyTriple<Integer, Integer, Float> triple : res1) {
+    for (MyTriple<Integer, Integer, Double> triple : res1) {
       apresentar.add(i - 1, new MyFour<>(Month.of(i), triple));
       i++;
     }
@@ -92,22 +92,22 @@ public class Query3View implements IQueryViewFX {
     panel.getChildren().add(new Label("Query 3"));
     panel.getChildren().add(new Label("Query Time: " + time));
 
-    TableView<MyFour<Month, Integer, Integer, Float>> tableView = new TableView<>();
-    TableColumn<MyFour<Month, Integer, Integer, Float>, String> mesesColumn = new TableColumn<>("Mes");
+    TableView<MyFour<Month, Integer, Integer, Double>> tableView = new TableView<>();
+    TableColumn<MyFour<Month, Integer, Integer, Double>, String> mesesColumn = new TableColumn<>("Mes");
     mesesColumn.setCellValueFactory(new PropertyValueFactory<>("first"));
     tableView.getColumns().add(mesesColumn);
 
-    TableColumn<MyFour<Month, Integer, Integer, Float>, String> column2 = new TableColumn<>("Number reviews total");
+    TableColumn<MyFour<Month, Integer, Integer, Double>, String> column2 = new TableColumn<>("Number reviews total");
     column2.setCellValueFactory(new PropertyValueFactory<>("left"));
 
     tableView.getColumns().add(column2);
 
-    TableColumn<MyFour<Month, Integer, Integer, Float>, String> column3 = new TableColumn<>("Number business distinct");
+    TableColumn<MyFour<Month, Integer, Integer, Double>, String> column3 = new TableColumn<>("Number business distinct");
     column3.setCellValueFactory(new PropertyValueFactory<>("middle"));
 
     tableView.getColumns().add(column3);
 
-    TableColumn<MyFour<Month, Integer, Integer, Float>, String> meanIdColumn = new TableColumn<>("Average");
+    TableColumn<MyFour<Month, Integer, Integer, Double>, String> meanIdColumn = new TableColumn<>("Average");
     meanIdColumn.setCellValueFactory(new PropertyValueFactory<>("right"));
     tableView.getColumns().add(meanIdColumn);
